Python
Given an array of integers with size 'num' and another number 'num2. Also given a size 'S' (integer value) of the subset. The task is to find if the number 'num2' is presentin every non-overlapping subset of size 'S' in the given array. If its present in every subset print 1 otherwise print 0.
For Example:
Input
12
20 1 6 8 13 20 17 20 13 16 20
20
3
Output
1